requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asqlite==2.0.0
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-HTTPAuth==4.8.0
Flask-SocketIO==5.5.1
frozenlist==1.4.1
h11==0.14.0
idna==3.10
iso8601==2.1.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
nodeenv==1.9.1
pyright==1.1.393
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
ruff==0.9.6
simple-websocket==1.1.0
twitchio==3.0.0b3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.13.1
